store projects in local storage
